Allow non-admins to tool factory if specific enviornment variable set.
Specifically GALAXY_DEVELOPMENT_ENVIRONENT must be set exactly to "1". I have a new `planemo tool_factory` command and it sets this environment variable to indicate that this is planemo running the Galaxy instances and therefore really should not be some sort of development server - and hence the tool factory should just work for whatever user.